Add new code action specification Start splitting up DataTypesJSON into smaller modules. Needed to do this since ClientCapabilities now references CodeAction types, which then led into a whole chain of module dependencies
