+++ haskell-lsp-types/src/Language/Haskell/LSP/TH/DataTypesJSON.hs
@@ -10,7 +10,18 @@
 {-# LANGUAGE TemplateHaskell            #-}
 {-# LANGUAGE TypeSynonymInstances       #-}
 
-module Language.Haskell.LSP.TH.DataTypesJSON where
+module Language.Haskell.LSP.TH.DataTypesJSON
+    ( module Language.Haskell.LSP.TH.DataTypesJSON
+    , module Language.Haskell.LSP.TH.CodeAction
+    , module Language.Haskell.LSP.TH.Command
+    , module Language.Haskell.LSP.TH.Diagnostic
+    , module Language.Haskell.LSP.TH.List
+    , module Language.Haskell.LSP.TH.Location
+    , module Language.Haskell.LSP.TH.Message
+    , module Language.Haskell.LSP.TH.TextDocumentIdentifier
+    , module Language.Haskell.LSP.TH.Uri
+    , module Language.Haskell.LSP.TH.WorkspaceEdit
+    ) where
 
 import           Control.Applicative
 import           Control.Lens.TH                            (makeFieldsNoPrefix)
@@ -22,924 +33,23 @@ import qualified Data.HashMap.Strict                        as H
 import           Data.Monoid                                ((<>))
 import           Data.Text                                  (Text)
 import qualified Data.Text                                  as T
-import           Network.URI
-import qualified System.FilePath.Posix                      as FPP
-import qualified System.FilePath.Windows                    as FPW
-import qualified System.Info
 import           System.IO                                  (FilePath)
 
 import           Language.Haskell.LSP.TH.ClientCapabilities
+import           Language.Haskell.LSP.TH.CodeAction
+import           Language.Haskell.LSP.TH.Command
 import           Language.Haskell.LSP.TH.Constants
+import           Language.Haskell.LSP.TH.Diagnostic
+import           Language.Haskell.LSP.TH.List
+import           Language.Haskell.LSP.TH.Location
+import           Language.Haskell.LSP.TH.Message
+import           Language.Haskell.LSP.TH.TextDocumentIdentifier
 import           Language.Haskell.LSP.TH.Utils
+import           Language.Haskell.LSP.TH.Uri
+import           Language.Haskell.LSP.TH.WorkspaceEdit
 
 -- ---------------------------------------------------------------------
 
--- | This data type is used to host a FromJSON instance for the encoding used by
--- elisp, where an empty list shows up as "null"
-newtype List a = List [a]
-                deriving (Show,Read,Eq,Ord,Monoid,Functor,Foldable,Traversable)
-
-instance (A.ToJSON a) => A.ToJSON (List a) where
-  toJSON (List ls) = toJSON ls
-
-instance (A.FromJSON a) => A.FromJSON (List a) where
-  parseJSON A.Null = return (List [])
-  parseJSON v      = List <$> parseJSON v
-
-#if __GLASGOW_HASKELL__ >= 804
-instance Semigroup (List a) where
-  (<>) = mappend
-#endif
-
--- ---------------------------------------------------------------------
-
-newtype Uri = Uri { getUri :: Text }
-  deriving (Eq,Ord,Read,Show,A.FromJSON,A.ToJSON,Hashable,A.ToJSONKey,A.FromJSONKey)
-
-fileScheme :: String
-fileScheme = "file:"
-
-windowsOS :: String
-windowsOS = "mingw32"
-
-type SystemOS = String
-
-uriToFilePath :: Uri -> Maybe FilePath
-uriToFilePath = platformAwareUriToFilePath System.Info.os
-
-platformAwareUriToFilePath :: String -> Uri -> Maybe FilePath
-platformAwareUriToFilePath systemOS (Uri uri) = do
-  parsedUri <- parseURI $ T.unpack uri
-  if uriScheme parsedUri == fileScheme
-    then return $ (platformAdjustFromUriPath systemOS . unEscapeString . uriPath) parsedUri
-    else Nothing
-
-platformAdjustFromUriPath :: SystemOS -> String -> FilePath
-platformAdjustFromUriPath systemOS srcPath =
-  if systemOS /= windowsOS then srcPath
-    else let
-      firstSegment:rest = (FPP.splitDirectories . tail) srcPath  -- Drop leading '/' for absolute Windows paths
-      drive = if FPW.isDrive firstSegment then FPW.addTrailingPathSeparator firstSegment else firstSegment
-      in FPW.joinDrive drive $ FPW.joinPath rest
-
-filePathToUri :: FilePath -> Uri
-filePathToUri = platformAwareFilePathToUri System.Info.os
-
-platformAwareFilePathToUri :: SystemOS -> FilePath -> Uri
-platformAwareFilePathToUri systemOS fp = Uri . T.pack . show $ URI
-  { uriScheme = fileScheme
-  , uriAuthority = Just $ URIAuth "" "" ""
-  , uriPath = platformAdjustToUriPath systemOS fp
-  , uriQuery = ""
-  , uriFragment = ""
-  }
-
-platformAdjustToUriPath :: SystemOS -> FilePath -> String
-platformAdjustToUriPath systemOS srcPath =
-  if systemOS /= windowsOS then srcPath
-    else let
-      drive:rest = FPW.splitDirectories srcPath
-      leaveCharUnescaped = (/= ':')
-      removePathSeparator = filter (not . FPW.isPathSeparator)
-      escapedDrive = removePathSeparator $ escapeURIString leaveCharUnescaped drive
-      in '/' : FPP.joinPath (escapedDrive : rest)
-
--- ---------------------------------------------------------------------
-
--- | Id used for a request, Can be either a String or an Int
-data LspId = IdInt Int | IdString Text
-            deriving (Show,Read,Eq,Ord)
-
-instance A.ToJSON LspId where
-  toJSON (IdInt i)    = toJSON i
-  toJSON (IdString s) = toJSON s
-
-instance A.FromJSON LspId where
-  parseJSON v@(A.Number _) = IdInt <$> parseJSON v
-  parseJSON  (A.String  s) = return (IdString s)
-  parseJSON _              = mempty
-
-instance Hashable LspId where
-  hashWithSalt salt (IdInt i) = hashWithSalt salt i
-  hashWithSalt salt (IdString s) = hashWithSalt salt s
-
--- ---------------------------------------------------------------------
-
--- | Id used for a response, Can be either a String or an Int, or Null. If a
--- request doesn't provide a result value the receiver of a request still needs
--- to return a response message to conform to the JSON RPC specification. The
--- result property of the ResponseMessage should be set to null in this case to
--- signal a successful request.
-data LspIdRsp = IdRspInt Int | IdRspString Text | IdRspNull
-            deriving (Show,Read,Eq)
-
-instance A.ToJSON LspIdRsp where
-  toJSON (IdRspInt i)    = toJSON i
-  toJSON (IdRspString s) = toJSON s
-  toJSON IdRspNull       = A.Null
-
-instance A.FromJSON LspIdRsp where
-  parseJSON v@(A.Number _) = IdRspInt <$> parseJSON v
-  parseJSON  (A.String  s) = return $ IdRspString s
-  parseJSON  A.Null        = return IdRspNull
-  parseJSON _              = mempty
-
-instance Hashable LspIdRsp where
-  hashWithSalt salt (IdRspInt i) = hashWithSalt salt i
-  hashWithSalt salt (IdRspString s) = hashWithSalt salt s
-  hashWithSalt _ IdRspNull = 0
-
--- | Converts an LspId to its LspIdRsp counterpart.
-responseId :: LspId -> LspIdRsp
-responseId (IdInt    i) = IdRspInt i
-responseId (IdString s) = IdRspString s
-
--- | Converts an LspIdRsp to its LspId counterpart. 
-requestId :: LspIdRsp -> LspId
-requestId (IdRspInt    i) = IdInt i
-requestId (IdRspString s) = IdString s
-requestId IdRspNull       = error "Null response id"
-
--- ---------------------------------------------------------------------
-
--- Client Methods
-data ClientMethod =
- -- General
-   Initialize
- | Initialized
- | Shutdown
- | Exit
- | CancelRequest
- -- Workspace
- | WorkspaceDidChangeConfiguration
- | WorkspaceDidChangeWatchedFiles
- | WorkspaceSymbol
- | WorkspaceExecuteCommand
- -- Document
- | TextDocumentDidOpen
- | TextDocumentDidChange
- | TextDocumentWillSave
- | TextDocumentWillSaveWaitUntil
- | TextDocumentDidSave
- | TextDocumentDidClose
- | TextDocumentCompletion
- | CompletionItemResolve
- | TextDocumentHover
- | TextDocumentSignatureHelp
- | TextDocumentReferences
- | TextDocumentDocumentHighlight
- | TextDocumentDocumentSymbol
- | TextDocumentFormatting
- | TextDocumentRangeFormatting
- | TextDocumentOnTypeFormatting
- | TextDocumentDefinition
- | TextDocumentCodeAction
- | TextDocumentCodeLens
- | CodeLensResolve
- | TextDocumentDocumentLink
- | DocumentLinkResolve
- | TextDocumentRename
- -- Messages of the form $/message
- -- Implementation Dependent, can be ignored
- | Misc Text
-   deriving (Eq,Ord,Read,Show)
-
-instance A.FromJSON ClientMethod where
-  -- General
-  parseJSON (A.String "initialize")                       = return Initialize
-  parseJSON (A.String "initialized")                      = return Initialized
-  parseJSON (A.String "shutdown")                         = return Shutdown
-  parseJSON (A.String "exit")                             = return Exit
-  parseJSON (A.String "$/cancelRequest")                  = return CancelRequest
- -- Workspace
-  parseJSON (A.String "workspace/didChangeConfiguration") = return WorkspaceDidChangeConfiguration
-  parseJSON (A.String "workspace/didChangeWatchedFiles")  = return WorkspaceDidChangeWatchedFiles
-  parseJSON (A.String "workspace/symbol")                 = return WorkspaceSymbol
-  parseJSON (A.String "workspace/executeCommand")         = return WorkspaceExecuteCommand
- -- Document
-  parseJSON (A.String "textDocument/didOpen")             = return TextDocumentDidOpen
-  parseJSON (A.String "textDocument/didChange")           = return TextDocumentDidChange
-  parseJSON (A.String "textDocument/willSave")            = return TextDocumentWillSave
-  parseJSON (A.String "textDocument/willSaveWaitUntil")   = return TextDocumentWillSaveWaitUntil
-  parseJSON (A.String "textDocument/didSave")             = return TextDocumentDidSave
-  parseJSON (A.String "textDocument/didClose")            = return TextDocumentDidClose
-  parseJSON (A.String "textDocument/completion")          = return TextDocumentCompletion
-  parseJSON (A.String "completionItem/resolve")           = return CompletionItemResolve
-  parseJSON (A.String "textDocument/hover")               = return TextDocumentHover
-  parseJSON (A.String "textDocument/signatureHelp")       = return TextDocumentSignatureHelp
-  parseJSON (A.String "textDocument/references")          = return TextDocumentReferences
-  parseJSON (A.String "textDocument/documentHighlight")   = return TextDocumentDocumentHighlight
-  parseJSON (A.String "textDocument/documentSymbol")      = return TextDocumentDocumentSymbol
-  parseJSON (A.String "textDocument/formatting")          = return TextDocumentFormatting
-  parseJSON (A.String "textDocument/rangeFormatting")     = return TextDocumentRangeFormatting
-  parseJSON (A.String "textDocument/onTypeFormatting")    = return TextDocumentOnTypeFormatting
-  parseJSON (A.String "textDocument/definition")          = return TextDocumentDefinition
-  parseJSON (A.String "textDocument/codeAction")          = return TextDocumentCodeAction
-  parseJSON (A.String "textDocument/codeLens")            = return TextDocumentCodeLens
-  parseJSON (A.String "codeLens/resolve")                 = return CodeLensResolve
-  parseJSON (A.String "textDocument/documentLink")        = return TextDocumentDocumentLink
-  parseJSON (A.String "documentLink/resolve")             = return DocumentLinkResolve
-  parseJSON (A.String "textDocument/rename")              = return TextDocumentRename
-  parseJSON (A.String x)                                  = if T.isPrefixOf "$/" x
-                                                               then return $ Misc (T.drop 2 x)
-                                                            else mempty
-  parseJSON _                                             = mempty
-
-instance A.ToJSON ClientMethod where
-  -- General
-  toJSON Initialize                      = A.String "initialize"
-  toJSON Initialized                     = A.String "initialized"
-  toJSON Shutdown                        = A.String "shutdown"
-  toJSON Exit                            = A.String "exit"
-  toJSON CancelRequest                   = A.String "$/cancelRequest"
-  -- Workspace
-  toJSON WorkspaceDidChangeConfiguration = A.String "workspace/didChangeConfiguration"
-  toJSON WorkspaceDidChangeWatchedFiles  = A.String "workspace/didChangeWatchedFiles"
-  toJSON WorkspaceSymbol                 = A.String "workspace/symbol"
-  toJSON WorkspaceExecuteCommand         = A.String "workspace/executeCommand"
-  -- Document
-  toJSON TextDocumentDidOpen             = A.String "textDocument/didOpen"
-  toJSON TextDocumentDidChange           = A.String "textDocument/didChange"
-  toJSON TextDocumentWillSave            = A.String "textDocument/willSave"
-  toJSON TextDocumentWillSaveWaitUntil   = A.String "textDocument/willSaveWaitUntil"
-  toJSON TextDocumentDidSave             = A.String "textDocument/didSave"
-  toJSON TextDocumentDidClose            = A.String "textDocument/didClose"
-  toJSON TextDocumentCompletion          = A.String "textDocument/completion"
-  toJSON CompletionItemResolve           = A.String "completionItem/resolve"
-  toJSON TextDocumentHover               = A.String "textDocument/hover"
-  toJSON TextDocumentSignatureHelp       = A.String "textDocument/signatureHelp"
-  toJSON TextDocumentReferences          = A.String "textDocument/references"
-  toJSON TextDocumentDocumentHighlight   = A.String "textDocument/documentHighlight"
-  toJSON TextDocumentDocumentSymbol      = A.String "textDocument/documentSymbol"
-  toJSON TextDocumentFormatting          = A.String "textDocument/formatting"
-  toJSON TextDocumentRangeFormatting     = A.String "textDocument/rangeFormatting"
-  toJSON TextDocumentOnTypeFormatting    = A.String "textDocument/onTypeFormatting"
-  toJSON TextDocumentDefinition          = A.String "textDocument/definition"
-  toJSON TextDocumentCodeAction          = A.String "textDocument/codeAction"
-  toJSON TextDocumentCodeLens            = A.String "textDocument/codeLens"
-  toJSON CodeLensResolve                 = A.String "codeLens/resolve"
-  toJSON TextDocumentRename              = A.String "textDocument/rename"
-  toJSON TextDocumentDocumentLink        = A.String "textDocument/documentLink"
-  toJSON DocumentLinkResolve             = A.String "documentLink/resolve"
-  toJSON (Misc xs)                       = A.String $ "$/" <> xs
-
-data ServerMethod =
-  -- Window
-    WindowShowMessage
-  | WindowShowMessageRequest
-  | WindowLogMessage
-  | TelemetryEvent
-  -- Client
-  | ClientRegisterCapability
-  | ClientUnregisterCapability
-  -- Workspace
-  | WorkspaceApplyEdit
-  -- Document
-  | TextDocumentPublishDiagnostics
-  -- Cancelling
-  | CancelRequestServer  
-   deriving (Eq,Ord,Read,Show)
-
-instance A.FromJSON ServerMethod where
-  -- Window
-  parseJSON (A.String "window/showMessage")              = return WindowShowMessage
-  parseJSON (A.String "window/showMessageRequest")       = return WindowShowMessageRequest
-  parseJSON (A.String "window/logMessage")               = return WindowLogMessage
-  parseJSON (A.String "telemetry/event")                 = return TelemetryEvent
-  -- Client
-  parseJSON (A.String "client/registerCapability")       = return ClientRegisterCapability
-  parseJSON (A.String "client/unregisterCapability")     = return ClientUnregisterCapability
-  -- Workspace
-  parseJSON (A.String "workspace/applyEdit")             = return WorkspaceApplyEdit
-  -- Document
-  parseJSON (A.String "textDocument/publishDiagnostics") = return TextDocumentPublishDiagnostics
-  -- Cancelling
-  parseJSON (A.String "$/cancelRequest")                 = return CancelRequestServer
-  parseJSON _                                            = mempty
-
-instance A.ToJSON ServerMethod where
-  -- Window
-  toJSON WindowShowMessage = A.String "window/showMessage"
-  toJSON WindowShowMessageRequest = A.String "window/showMessageRequest"
-  toJSON WindowLogMessage = A.String "window/logMessage"
-  toJSON TelemetryEvent = A.String "telemetry/event"
-  -- Client
-  toJSON ClientRegisterCapability = A.String "client/registerCapability"
-  toJSON ClientUnregisterCapability = A.String "client/unregisterCapability"
-  -- Workspace
-  toJSON WorkspaceApplyEdit = A.String "workspace/applyEdit"
-  -- Document
-  toJSON TextDocumentPublishDiagnostics = A.String "textDocument/publishDiagnostics"
-  -- Cancelling
-  toJSON CancelRequestServer = A.String "$/cancelRequest"
-
-data RequestMessage m req resp =
-  RequestMessage
-    { _jsonrpc :: Text
-    , _id      :: LspId
-    , _method  :: m
-    , _params  :: req
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''RequestMessage
-makeFieldsNoPrefix ''RequestMessage
-
--- ---------------------------------------------------------------------
-{-
-interface ResponseError<D> {
-    /**
-     * A number indicating the error type that occurred.
-     */
-    code: number;
-
-    /**
-     * A string providing a short description of the error.
-     */
-    message: string;
-
-    /**
-     * A Primitive or Structured value that contains additional
-     * information about the error. Can be omitted.
-     */
-    data?: D;
-}
-
-export namespace ErrorCodes {
-        // Defined by JSON RPC
-        export const ParseError: number = -32700;
-        export const InvalidRequest: number = -32600;
-        export const MethodNotFound: number = -32601;
-        export const InvalidParams: number = -32602;
-        export const InternalError: number = -32603;
-        export const serverErrorStart: number = -32099;
-        export const serverErrorEnd: number = -32000;
-        export const ServerNotInitialized: number = -32002;
-        export const UnknownErrorCode: number = -32001;
-
-        // Defined by the protocol.
-        export const RequestCancelled: number = -32800;
-}
--}
-
-data ErrorCode = ParseError
-               | InvalidRequest
-               | MethodNotFound
-               | InvalidParams
-               | InternalError
-               | ServerErrorStart
-               | ServerErrorEnd
-               | ServerNotInitialized
-               | UnknownErrorCode
-               | RequestCancelled
-               -- ^ Note: server error codes are reserved from -32099 to -32000
-               deriving (Read,Show,Eq)
-
-instance A.ToJSON ErrorCode where
-  toJSON ParseError           = A.Number (-32700)
-  toJSON InvalidRequest       = A.Number (-32600)
-  toJSON MethodNotFound       = A.Number (-32601)
-  toJSON InvalidParams        = A.Number (-32602)
-  toJSON InternalError        = A.Number (-32603)
-  toJSON ServerErrorStart     = A.Number (-32099)
-  toJSON ServerErrorEnd       = A.Number (-32000)
-  toJSON ServerNotInitialized = A.Number (-32002)
-  toJSON UnknownErrorCode     = A.Number (-32001)
-  toJSON RequestCancelled     = A.Number (-32800)
-
-instance A.FromJSON ErrorCode where
-  parseJSON (A.Number (-32700)) = pure ParseError
-  parseJSON (A.Number (-32600)) = pure InvalidRequest
-  parseJSON (A.Number (-32601)) = pure MethodNotFound
-  parseJSON (A.Number (-32602)) = pure InvalidParams
-  parseJSON (A.Number (-32603)) = pure InternalError
-  parseJSON (A.Number (-32099)) = pure ServerErrorStart
-  parseJSON (A.Number (-32000)) = pure ServerErrorEnd
-  parseJSON (A.Number (-32002)) = pure ServerNotInitialized
-  parseJSON (A.Number (-32001)) = pure UnknownErrorCode
-  parseJSON (A.Number (-32800)) = pure RequestCancelled
-  parseJSON _                   = mempty
-
--- -------------------------------------
-
-data ResponseError =
-  ResponseError
-    { _code    :: ErrorCode
-    , _message :: Text
-    , _xdata   :: Maybe A.Value
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions{ fieldLabelModifier = customModifier } ''ResponseError
-makeFieldsNoPrefix ''ResponseError
-
--- ---------------------------------------------------------------------
-
-data ResponseMessage a =
-  ResponseMessage
-    { _jsonrpc :: Text
-    , _id      :: LspIdRsp
-    , _result  :: Maybe a
-    , _error   :: Maybe ResponseError
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''ResponseMessage
-makeFieldsNoPrefix ''ResponseMessage
-
-type ErrorResponse = ResponseMessage ()
-
--- ---------------------------------------------------------------------
-
-type BareResponseMessage = ResponseMessage A.Value
-
--- ---------------------------------------------------------------------
-{-
-$ Notifications and Requests
-
-Notification and requests ids starting with '$/' are messages which are protocol
-implementation dependent and might not be implementable in all clients or
-servers. For example if the server implementation uses a single threaded
-synchronous programming language then there is little a server can do to react
-to a '$/cancelRequest'. If a server or client receives notifications or requests
-starting with '$/' it is free to ignore them if they are unknown.
--}
-
-data NotificationMessage m a =
-  NotificationMessage
-    { _jsonrpc :: Text
-    , _method  :: m
-    , _params  :: a
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''NotificationMessage
-makeFieldsNoPrefix ''NotificationMessage
-
--- ---------------------------------------------------------------------
-{-
-Cancellation Support
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#cancellation-support
-
-    New: The base protocol now offers support for request cancellation. To
-    cancel a request, a notification message with the following properties is
-    sent:
-
-Notification:
-
-    method: '$/cancelRequest'
-    params: CancelParams defined as follows:
-
-interface CancelParams {
-    /**
-     * The request id to cancel.
-     */
-    id: number | string;
-}
-
-A request that got canceled still needs to return from the server and send a
-response back. It can not be left open / hanging. This is in line with the JSON
-RPC protocol that requires that every request sends a response back. In addition
-it allows for returning partial results on cancel.
--}
-
-data CancelParams =
-  CancelParams
-    { _id :: LspId
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''CancelParams
-makeFieldsNoPrefix ''CancelParams
-
-type CancelNotification = NotificationMessage ClientMethod CancelParams
-type CancelNotificationServer = NotificationMessage ServerMethod CancelParams
-
--- ---------------------------------------------------------------------
-
-{-
-The current protocol is talored for textual documents which content can be
-represented as a string. There is currently no support for binary documents.
-Positions inside a document (see Position definition below) are expressed as a
-zero-based line and character offset. To ensure that both client and server
-split the string into the same line representation the protocol specs the
-following end of line sequences: '\n', '\r\n' and '\r'.
-
-export const EOL: string[] = ['\n', '\r\n', '\r'];
--}
-{-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#position
-
-Position in a text document expressed as zero-based line and character offset. A
-position is between two characters like an 'insert' cursor in a editor.
-
-interface Position {
-    /**
-     * Line position in a document (zero-based).
-     */
-    line: number;
-
-    /**
-     * Character offset on a line in a document (zero-based).
-     */
-    character: number;
-}
--}
-data Position =
-  Position
-    { _line      :: Int
-    , _character :: Int
-    } deriving (Show, Read, Eq, Ord)
-
-deriveJSON lspOptions ''Position
-makeFieldsNoPrefix ''Position
-
--- ---------------------------------------------------------------------
-{-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#range
-
-A range in a text document expressed as (zero-based) start and end positions. A
-range is comparable to a selection in an editor. Therefore the end position is
-exclusive.
-
-interface Range {
-    /**
-     * The range's start position.
-     */
-    start: Position;
-
-    /**
-     * The range's end position.
-     */
-    end: Position;
-}
--}
-
-data Range =
-  Range
-    { _start :: Position
-    , _end   :: Position
-    } deriving (Show, Read, Eq, Ord)
-
-deriveJSON lspOptions ''Range
-makeFieldsNoPrefix ''Range
-
--- ---------------------------------------------------------------------
-{-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#location
-
-Represents a location inside a resource, such as a line inside a text file.
-
-interface Location {
-    uri: string;
-    range: Range;
-}
--}
-
-data Location =
-  Location
-    { _uri   :: Uri
-    , _range :: Range
-    } deriving (Show, Read, Eq, Ord)
-
-deriveJSON lspOptions ''Location
-makeFieldsNoPrefix ''Location
-
--- ---------------------------------------------------------------------
-{-
-The protocol currently supports the following diagnostic severities:
-
-enum DiagnosticSeverity {
-    /**
-     * Reports an error.
-     */
-    Error = 1,
-    /**
-     * Reports a warning.
-     */
-    Warning = 2,
-    /**
-     * Reports an information.
-     */
-    Information = 3,
-    /**
-     * Reports a hint.
-     */
-    Hint = 4
-}
--}
-data DiagnosticSeverity
-  = DsError   -- ^ Error = 1,
-  | DsWarning -- ^ Warning = 2,
-  | DsInfo    -- ^ Info = 3,
-  | DsHint    -- ^ Hint = 4
-  deriving (Eq,Ord,Show,Read)
-
-instance A.ToJSON DiagnosticSeverity where
-  toJSON DsError   = A.Number 1
-  toJSON DsWarning = A.Number 2
-  toJSON DsInfo    = A.Number 3
-  toJSON DsHint    = A.Number 4
-
-instance A.FromJSON DiagnosticSeverity where
-  parseJSON (A.Number 1) = pure DsError
-  parseJSON (A.Number 2) = pure DsWarning
-  parseJSON (A.Number 3) = pure DsInfo
-  parseJSON (A.Number 4) = pure DsHint
-  parseJSON _            = mempty
-
--- ---------------------------------------------------------------------
-{-
-Represents a related message and source code location for a diagnostic. This should be
-used to point to code locations that cause or related to a diagnostics, e.g when duplicating
-a symbol in a scope.
-
-export interface DiagnosticRelatedInformation {
-  /**
-   * The location of this related diagnostic information.
-   */
-  location: Location;
-
-  /**
-   * The message of this related diagnostic information.
-   */
-  message: string;
-}
--}
-
-data DiagnosticRelatedInformation =
-  DiagnosticRelatedInformation
-    { _location :: Location
-    , _message  :: Text
-    } deriving (Show, Read, Eq, Ord)
-
-deriveJSON lspOptions ''DiagnosticRelatedInformation
-makeFieldsNoPrefix ''DiagnosticRelatedInformation
-
--- ---------------------------------------------------------------------
-{-
-Diagnostic
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#diagnostic
-
-Represents a diagnostic, such as a compiler error or warning. Diagnostic objects
-are only valid in the scope of a resource.
-
-interface Diagnostic {
-    /**
-     * The range at which the message applies.
-     */
-    range: Range;
-
-    /**
-     * The diagnostic's severity. Can be omitted. If omitted it is up to the
-     * client to interpret diagnostics as error, warning, info or hint.
-     */
-    severity?: number;
-
-    /**
-     * The diagnostic's code. Can be omitted.
-     */
-    code?: number | string;
-
-    /**
-     * A human-readable string describing the source of this
-     * diagnostic, e.g. 'typescript' or 'super lint'.
-     */
-    source?: string;
-
-    /**
-     * The diagnostic's message.
-     */
-    message: string;
-
-    /**
-     * An array of related diagnostic information, e.g. when symbol-names within
-     * a scope collide all definitions can be marked via this property.
-     */
-    relatedInformation?: DiagnosticRelatedInformation[];
-}
--}
-
-type DiagnosticSource = Text
-data Diagnostic =
-  Diagnostic
-    { _range              :: Range
-    , _severity           :: Maybe DiagnosticSeverity
-    , _code               :: Maybe Text -- Note: Protocol allows Int too.
-    , _source             :: Maybe DiagnosticSource
-    , _message            :: Text
-    , _relatedInformation :: Maybe (List DiagnosticRelatedInformation)
-    } deriving (Show, Read, Eq, Ord)
-
-deriveJSON lspOptions ''Diagnostic
-makeFieldsNoPrefix ''Diagnostic
-
--- ---------------------------------------------------------------------
-{-
-Command
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#command
-
-Represents a reference to a command. Provides a title which will be used to
-represent a command in the UI. Commands are identitifed using a string
-identifier and the protocol currently doesn't specify a set of well known
-commands. So executing a command requires some tool extension code.
-
-interface Command {
-    /**
-     * Title of the command, like `save`.
-     */
-    title: string;
-    /**
-     * The identifier of the actual command handler.
-     */
-    command: string;
-    /**
-     * Arguments that the command handler should be
-     * invoked with.
-     */
-    arguments?: any[];
-}
--}
-
-data Command =
-  Command
-    { _title     :: Text
-    , _command   :: Text
-    , _arguments :: Maybe A.Value
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''Command
-makeFieldsNoPrefix ''Command
-
--- ---------------------------------------------------------------------
-{-
-TextEdit
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textedit
-
-A textual edit applicable to a text document.
-
-interface TextEdit {
-    /**
-     * The range of the text document to be manipulated. To insert
-     * text into a document create a range where start === end.
-     */
-    range: Range;
-
-    /**
-     * The string to be inserted. For delete operations use an
-     * empty string.
-     */
-    newText: string;
-}
-
-
--}
-
-data TextEdit =
-  TextEdit
-    { _range   :: Range
-    , _newText :: Text
-    } deriving (Show,Read,Eq)
-
-deriveJSON lspOptions ''TextEdit
-makeFieldsNoPrefix ''TextEdit
-
--- ---------------------------------------------------------------------
-{-
-VersionedTextDocumentIdentifier
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#versionedtextdocumentidentifier
-
-    New: An identifier to denote a specific version of a text document.
-
-interface VersionedTextDocumentIdentifier extends TextDocumentIdentifier {
-    /**
-     * The version number of this document.
-     */
-    version: number;
--}
-
-type TextDocumentVersion = Int
-
-data VersionedTextDocumentIdentifier =
-  VersionedTextDocumentIdentifier
-    { _uri     :: Uri
-    , _version :: TextDocumentVersion
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''VersionedTextDocumentIdentifier
-makeFieldsNoPrefix ''VersionedTextDocumentIdentifier
-
--- ---------------------------------------------------------------------
-{-
-New in 3.0
-----------
-
-TextDocumentEdit
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#new-textdocumentedit
-
-If multiple TextEdits are applied to a text document, all text edits describe
-changes made to the initial document version. Execution wise text edits should
-applied from the bottom to the top of the text document. Overlapping text edits
-are not supported.
-
-export interface TextDocumentEdit {
-        /**
-         * The text document to change.
-         */
-        textDocument: VersionedTextDocumentIdentifier;
-
-        /**
-         * The edits to be applied.
-         */
-        edits: TextEdit[];
-}
-
--}
-
-data TextDocumentEdit =
-  TextDocumentEdit
-    { _textDocument :: VersionedTextDocumentIdentifier
-    , _edits        :: List TextEdit
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''TextDocumentEdit
-makeFieldsNoPrefix ''TextDocumentEdit
-
--- ---------------------------------------------------------------------
-{-
-Changed in 3.0
---------------
-
-WorkspaceEdit
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#workspaceedit
-
-
-Changed A workspace edit represents changes to many resources managed in the
-workspace. The edit should either provide changes or documentChanges. If
-documentChanges are present they are preferred over changes if the client can
-handle versioned document edits.
-
-export interface WorkspaceEdit {
-        /**
-         * Holds changes to existing resources.
-         */
-        changes?: { [uri: string]: TextEdit[]; };
-
-        /**
-         * An array of `TextDocumentEdit`s to express changes to specific a specific
-         * version of a text document. Whether a client supports versioned document
-         * edits is expressed via `WorkspaceClientCapabilities.versionedWorkspaceEdit`.
-         */
-        documentChanges?: TextDocumentEdit[];
-}
--}
-
-type WorkspaceEditMap = H.HashMap Uri (List TextEdit)
-
-data WorkspaceEdit =
-  WorkspaceEdit
-    { _changes         :: Maybe WorkspaceEditMap
-    , _documentChanges :: Maybe (List TextDocumentEdit)
-    } deriving (Show, Read, Eq)
-
-instance Monoid WorkspaceEdit where
-  mempty = WorkspaceEdit Nothing Nothing
-  mappend (WorkspaceEdit a b) (WorkspaceEdit c d) = WorkspaceEdit (a <> c) (b <> d)
-
-deriveJSON lspOptions ''WorkspaceEdit
-makeFieldsNoPrefix ''WorkspaceEdit
-
-#if __GLASGOW_HASKELL__ >= 804
-instance Semigroup WorkspaceEdit where
-  (<>) = mappend
-#endif
-
--- ---------------------------------------------------------------------
-{-
-TextDocumentIdentifier
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textdocumentidentifier
-
-Text documents are identified using a URI. On the protocol level, URIs are
-passed as strings. The corresponding JSON structure looks like this:
-
-interface TextDocumentIdentifier {
-    /**
-     * The text document's URI.
-     */
-    uri: string;
-}
--}
-data TextDocumentIdentifier =
-  TextDocumentIdentifier
-    { _uri :: Uri
-    } deriving (Show, Read, Eq)
-
-deriveJSON lspOptions ''TextDocumentIdentifier
-makeFieldsNoPrefix ''TextDocumentIdentifier
-
--- ---------------------------------------------------------------------
 
 {-
 TextDocumentItem
@@ -3601,84 +2711,6 @@ type WorkspaceSymbolsResponse = ResponseMessage (List SymbolInformation)
 
 -- ---------------------------------------------------------------------
 {-
-Code Action Request
-
-https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#code-action-request
-
-The code action request is sent from the client to the server to compute
-commands for a given text document and range. The request is triggered when the
-user moves the cursor into a problem marker in the editor or presses the
-lightbulb associated with a marker.
-
-Request
-
-    method: 'textDocument/codeAction'
-    params: CodeActionParams defined as follows:
-
-/**
- * Params for the CodeActionRequest
- */
-interface CodeActionParams {
-    /**
-     * The document in which the command was invoked.
-     */
-    textDocument: TextDocumentIdentifier;
-
-    /**
-     * The range for which the command was invoked.
-     */
-    range: Range;
-
-    /**
-     * Context carrying additional information.
-     */
-    context: CodeActionContext;
-}
-
-/**
- * Contains additional diagnostic information about the context in which
- * a code action is run.
- */
-interface CodeActionContext {
-    /**
-     * An array of diagnostics.
-     */
-    diagnostics: Diagnostic[];
-}
-
-Response
-
-    result: Command[] defined as follows:
-    error: code and message set in case an exception happens during the code
-           action request.
-
--}
-
-data CodeActionContext =
-  CodeActionContext
-    { _diagnostics :: List Diagnostic
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''CodeActionContext
-makeFieldsNoPrefix ''CodeActionContext
-
-
-data CodeActionParams =
-  CodeActionParams
-    { _textDocument :: TextDocumentIdentifier
-    , _range        :: Range
-    , _context      :: CodeActionContext
-    } deriving (Read,Show,Eq)
-
-deriveJSON lspOptions ''CodeActionParams
-makeFieldsNoPrefix ''CodeActionParams
-
-
-type CodeActionRequest  = RequestMessage ClientMethod CodeActionParams (List Command)
-type CodeActionResponse = ResponseMessage (List Command)
-
--- ---------------------------------------------------------------------
-{-
 Code Lens Request
 
 The code lens request is sent from the client to the server to compute code
@@ -4317,3 +3349,39 @@ makeFieldsNoPrefix ''TraceNotification
 
 
 -- ---------------------------------------------------------------------
+
+-- Generating lenses
+-- Need to do this all in one file or else we get duplicate functions
+
+-- Location
+makeFieldsNoPrefix ''Position
+makeFieldsNoPrefix ''Range
+makeFieldsNoPrefix ''Location
+
+-- CodeActions
+makeFieldsNoPrefix ''CodeActionContext
+makeFieldsNoPrefix ''CodeActionParams
+makeFieldsNoPrefix ''CodeAction
+
+-- WorkspaceEdit
+makeFieldsNoPrefix ''TextEdit
+makeFieldsNoPrefix ''VersionedTextDocumentIdentifier
+makeFieldsNoPrefix ''TextDocumentEdit
+makeFieldsNoPrefix ''WorkspaceEdit
+
+-- Message
+makeFieldsNoPrefix ''RequestMessage
+makeFieldsNoPrefix ''ResponseError
+makeFieldsNoPrefix ''ResponseMessage
+makeFieldsNoPrefix ''NotificationMessage
+makeFieldsNoPrefix ''CancelParams
+
+-- TextDocumentIdentifier
+makeFieldsNoPrefix ''TextDocumentIdentifier
+
+-- Command
+makeFieldsNoPrefix ''Command
+
+-- Diagnostic
+makeFieldsNoPrefix ''Diagnostic
+makeFieldsNoPrefix ''DiagnosticRelatedInformation
